BUG (PROPOSED FIX): Added comments to JotShortcut.h explaining why the `shortDescription` and `key` properties should not be defined as `assign`.
Changed `assign` to `copy` to suggest the desired change. (Obviously this has no effect on the compiled code in the static library, but, provided this change is back-ported to the original project source code, the static framework is re-compiled, and the binary copy is updated in this project, that should fix the crash exhibited in the code I added to ViewController).